fix/chore: initialize local services
