+++ plugin/src/main/kotlin/com/vaadin/plugin/listeners/ConfigurationCheckVaadinProjectListener.kt
@@ -22,6 +22,7 @@ import com.intellij.openapi.vcs.impl.ProjectLevelVcsManagerImpl
 import com.vaadin.plugin.actions.VaadinCompileOnSaveActionInfo
 import com.vaadin.plugin.copilot.CopilotPluginUtil
 import com.vaadin.plugin.copilot.service.CompilationStatusManagerService
+import com.vaadin.plugin.copilot.service.CopilotUndoManager
 import com.vaadin.plugin.utils.VaadinHomeUtil
 import com.vaadin.plugin.utils.VaadinIcons
 import com.vaadin.plugin.utils.trackPluginInitialized
@@ -52,12 +53,16 @@ class ConfigurationCheckVaadinProjectListener : VaadinProjectListener {
             checkVcsAddConfirmationSetting(project)
             checkCompileOnSave(project)
             initAmplitude(project)
+            initLocalServices(project)
             RunOnceUtil.runOnceForApp("hotswap-version-check-" + CopilotPluginUtil.getPluginVersion()) {
                 VaadinHomeUtil.updateOrInstallHotSwapJar()
             }
-            val compilationStatusManagerService = project.getService(CompilationStatusManagerService::class.java)
-            compilationStatusManagerService.subscribeToListenCompilationStatus()
         }
+    }
+
+    private fun initLocalServices(project: Project) {
+        project.getService(CompilationStatusManagerService::class.java).subscribeToCompilationStatus()
+        project.getService(CopilotUndoManager::class.java).subscribeToVfsChanges()
     }
 
     private fun checkReloadClassesSetting(project: Project) {
