+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -70,9 +70,6 @@
         <statusBarWidgetFactory implementation="com.vaadin.plugin.ui.VaadinStatusBarWidgetFactory"
                                 id="VaadinStatusBarWidgetFactory" order="last"/>
         <httpRequestHandler implementation="com.vaadin.plugin.copilot.service.CopilotRestService"/>
-        <projectService
-                serviceInterface="com.vaadin.plugin.copilot.service.CopilotUndoManager"
-                serviceImplementation="com.vaadin.plugin.copilot.service.CopilotUndoManagerImpl"/>
 
         <!-- Vaadin Project Wizard -->
         <moduleBuilder id="VaadinModuleBuilder" builderClass="com.vaadin.plugin.module.VaadinProjectBuilderAdapter"/>
